Module sst::ui

ui.rs

This is stuff related to the user interface (ui)

Modules

logo

logo.rs

misc

misc.rs

sensor

sensor.rs

Functions

abbrev

Check if what is an abbreviation of full and starts with least.

command_processor

command_processor

determine_cmd_type

determine_cmd_type

disp_title

disp_title

red_error

red_error

red_syntax_error

red_syntax_error